Move InvestmentSet to a new markets module
